+++ src/test/java/javax/time/calendar/format/TestTextParser.java
@@ -98,7 +98,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL);
         int newPos = pp.parse(context, "Monday", 0);
         assertEquals(newPos, 6);
-        assertEquals(context.getParsed(DAY_OF_WEEK), 1);
+        assertEquals(context.getParsed(DAY_OF_WEEK), (long) 1);
     }
 
     public void test_parse_midStr() throws Exception {
@@ -106,7 +106,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL);
         int newPos = pp.parse(context, "XxxMondayXxx", 3);
         assertEquals(newPos, 9);
-        assertEquals(context.getParsed(DAY_OF_WEEK), 1);
+        assertEquals(context.getParsed(DAY_OF_WEEK), (long) 1);
     }
 
     public void test_parse_remainderIgnored() throws Exception {
@@ -114,7 +114,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(DAY_OF_WEEK, TextStyle.SHORT);
         int newPos = pp.parse(context, "Wednesday", 0);
         assertEquals(newPos, 3);
-        assertEquals(context.getParsed(DAY_OF_WEEK), 3);
+        assertEquals(context.getParsed(DAY_OF_WEEK), (long) 3);
     }
 
     //-----------------------------------------------------------------------
@@ -191,7 +191,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(rule, style);
         int newPos = pp.parse(context, input, 0);
         assertEquals(newPos, input.length());
-        assertEquals(context.getParsed(rule), dow);
+        assertEquals(context.getParsed(rule), (long) dow);
     }
 
     @Test(dataProvider="parseNumber")
@@ -200,7 +200,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(rule, style);
         int newPos = pp.parse(context, input, 0);
         assertEquals(newPos, input.length());
-        assertEquals(context.getParsed(rule), dow);
+        assertEquals(context.getParsed(rule), (long) dow);
     }
 
     //-----------------------------------------------------------------------
@@ -221,7 +221,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(rule, style);
         int newPos = pp.parse(context, input.toUpperCase(), 0);
         assertEquals(newPos, input.length());
-        assertEquals(context.getParsed(rule), dow);
+        assertEquals(context.getParsed(rule), (long) dow);
     }
 
     //-----------------------------------------------------------------------
@@ -242,7 +242,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(rule, style);
         int newPos = pp.parse(context, input.toLowerCase(), 0);
         assertEquals(newPos, input.length());
-        assertEquals(context.getParsed(rule), dow);
+        assertEquals(context.getParsed(rule), (long) dow);
     }
 
     //-----------------------------------------------------------------------
@@ -254,7 +254,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL);
         int newPos = pp.parse(context, "January", 0);
         assertEquals(newPos, 7);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_full_strict_short_noMatch() throws Exception {
@@ -282,7 +282,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT);
         int newPos = pp.parse(context, "January", 0);
         assertEquals(newPos, 3);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_short_strict_short_match() throws Exception {
@@ -291,7 +291,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT);
         int newPos = pp.parse(context, "Janua", 0);
         assertEquals(newPos, 3);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_short_strict_number_noMatch() throws Exception {
@@ -321,7 +321,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT);
         int newPos = pp.parse(context, "janv.", 0);
         assertEquals(newPos, 5);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     //-----------------------------------------------------------------------
@@ -331,7 +331,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL);
         int newPos = pp.parse(context, "January", 0);
         assertEquals(newPos, 7);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_full_lenient_short_match() throws Exception {
@@ -340,7 +340,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL);
         int newPos = pp.parse(context, "Janua", 0);
         assertEquals(newPos, 3);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_full_lenient_number_match() throws Exception {
@@ -349,7 +349,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL);
         int newPos = pp.parse(context, "1", 0);
         assertEquals(newPos, 1);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     //-----------------------------------------------------------------------
@@ -359,7 +359,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT);
         int newPos = pp.parse(context, "January", 0);
         assertEquals(newPos, 7);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_short_lenient_short_match() throws Exception {
@@ -368,7 +368,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT);
         int newPos = pp.parse(context, "Janua", 0);
         assertEquals(newPos, 3);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
     public void test_parse_short_lenient_number_match() throws Exception {
@@ -377,7 +377,7 @@ public class TestTextParser {
         TextPrinterParser pp = new TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT);
         int newPos = pp.parse(context, "1", 0);
         assertEquals(newPos, 1);
-        assertEquals(context.getParsed(MONTH_OF_YEAR), 1);
+        assertEquals(context.getParsed(MONTH_OF_YEAR), (long) 1);
     }
 
 }
